requirements.txt
# Core AI / LLM
requests
openai

# Speech-to-Text (Whisper)
openai-whisper
sounddevice
numpy

# Text-to-Speech (Offline pyttsx3)
pyttsx3

# Screen Reader (OCR)
pytesseract
pillow

# Keyboard & Mouse automation
pyautogui
pynput

# Utils
rich
